Consolidated frame attribute assigning.
