+++ src/3p-frame.js
@@ -15,17 +15,14 @@
  */
 
 import {dev, user} from './log';
-import {isExperimentOn, experimentToggles, isCanary} from './experiments';
+import {isExperimentOn} from './experiments';
 import {getContextMetadata} from '../src/iframe-attributes';
 import {tryParseJson} from './json';
 import {getMode} from './mode';
-import {getModeObject} from './mode-object';
 import {dashToCamelCase} from './string';
 import {parseUrl, assertHttpsUrl} from './url';
-import {viewerForDoc} from './viewer';
 import {urls} from './config';
 import {setStyle} from './style';
-import {domFingerprint} from './utils/dom-fingerprint';
 
 /** @type {!Object<string,number>} Number of 3p frames on the for that type. */
 let count = {};
@@ -55,18 +52,7 @@ function getFrameAttributes(parentWindow, element, opt_type, opt_context) {
   attributes = getContextMetadata(parentWindow, element, sentinel,
       attributes);
   attributes.type = type;
-  const viewer = viewerForDoc(element);
-  const additionalContext = {
-    tagName: element.tagName,
-    mode: getModeObject(),
-    canary: isCanary(parentWindow),
-    hidden: !viewer.isVisible(),
-    initialIntersection: element.getIntersectionChangeEntry(),
-    domFingerprint: domFingerprint(element),
-    experimentToggles: experimentToggles(parentWindow),
-  };
   Object.assign(attributes._context, opt_context);
-  Object.assign(attributes._context, additionalContext);
   return attributes;
 }
