+++ src/iframe-attributes.js
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 import {documentInfoForDoc} from './document-info';
-import {isExperimentOn} from './experiments';
+import {isExperimentOn, experimentToggles, isCanary} from './experiments';
 import {viewerForDoc} from './viewer';
 import {getLengthNumeral} from './layout';
+import {getModeObject} from './mode-object';
+import {domFingerprint} from './utils/dom-fingerprint';
 
 /**
  * Produces the attributes for the ad template.
@@ -43,7 +45,8 @@ export function getContextMetadata(
   }
 
   const docInfo = documentInfoForDoc(element);
-  const referrer = viewerForDoc(element).getUnconfirmedReferrerUrl();
+  const viewer = viewerForDoc(element);
+  const referrer = viewer.getUnconfirmedReferrerUrl();
 
   const sentinelNameChange = isExperimentOn(
       parentWindow, 'sentinel-name-change');
@@ -57,6 +60,14 @@ export function getContextMetadata(
       href: locationHref,
     },
     startTime,
+    tagName: element.tagName,
+    mode: getModeObject(),
+    canary: isCanary(parentWindow),
+    hidden: !viewer.isVisible(),
+    initialIntersection: element.getIntersectionChangeEntry ?
+        element.getIntersectionChangeEntry() : null,
+    domFingerprint: domFingerprint(element),
+    experimentToggles: experimentToggles(parentWindow),
   };
   attributes._context[sentinelNameChange ? 'sentinel' : 'amp3pSentinel'] =
       sentinel;
